+++ app.vue
@@ -1,5 +1,7 @@
 <template>
   <div>
+    <LayoutHeader />
     <NuxtChild />
+    <LayoutFooter />
   </div>
 </template>
